requirements.txt
autopep8==2.0.1
importlib-metadata==5.2.0
importlib-resources==5.10.2
libusb==1.0.26b5
numpy==1.24.1
packaging==22.0
pkg-about==1.0.8
PyAudio==0.2.13
pycodestyle==2.10.0
pygame==2.1.3.dev8
pyusb==1.2.1
tomli==2.0.1
zipp==3.11.0
